fix: handle terminal sizing correctly
Created a wrapper around the CrosstermBackend to intercept calls to size
and window_size methods. Recreate the terminal whenever the window size
changes (there's probably a way to store the backend and update the size
without recreating the terminal, but this is simpler for now).

Also removed the terminal_sizes field from the Game struct, as it's no
longer needed (the players can have different terminal sizes).

Added more logs to the AppHandler methods to track the client ID and
terminal size changes.